Move structs for custom menu drawing to private header
so that they can be reused elsewhere.

No real changes, just prepare for the next commit.
